+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 Full documentatino for AMD Debugger API is available at
 [docs.amd.com](https://docs.amd.com/category/ROCDebugger%20API%20Guides).
 
+## (Unreleased) rocm-dbgapi-0.70.1
+### Added
+- Add support for gfx940, gfx941 and gfx942 architectures.
+
 ## (Unreleased) rocm-dbgapi-0.70.0
 ### Changed
 - The name reported for each agent is now based on the information stored